Make rotation reset work on devices that don't support rotation on home screen
